third_party/blink/web_tests/virtual/saa-beyond-cookies-disabled/external/wpt/storage-access-api/storage-access-beyond-cookies.blobStorage.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify StorageAccessAPIBeyondCookies for Blob Storage
  assert_equals: Storage Access API should be accessible and return first-party data expected "HasAccess for blobStorage" but got "Unable to load handle in same-origin context for blobStorage"
Harness: the test ran to completion.
